backend/requirements.txt
# Core
Flask>=2.3.3
Flask-CORS>=4.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.5
python-dotenv>=1.0.0
Werkzeug>=2.3.7

# Database
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.12.1

# File Processing
Pillow>=10.0.0
qrcode>=7.4.2

# Security
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
flask-limiter>=3.5.0

# API Documentation
flask-restx>=1.1.0
apispec>=6.3.1
marshmallow>=3.20.1

# Testing
pytest>=7.4.2
pytest-cov>=4.1.0
pytest-mock>=3.12.0
responses>=0.24.1

# Utilities
requests>=2.31.0
psutil>=5.9.5
python-dateutil>=2.8.2